requirements.txt
pywebview
pyserial
pyinstaller